backend_monitoring/requirements.txt
asgiref==3.5.2
asyncio==3.4.3
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
backports.zoneinfo;python_version<"3.9"
beautifulsoup4==4.11.1
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==38.0.4
daphne==4.0.0
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pika==1.3.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pytz==2022.6
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
zope.interface==5.5.2
